Fix coding according to code review
